# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Clinic API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  qualification: string;
  /** Years of experience */
  experience: number;
  /** Consultation fee in PKR */
  fee: number;
  availableTimes: string[];
  rating: number;
  totalPatients: number;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  upcoming: "upcoming",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Appointment {
  id: number;
  doctorId: number;
  doctorName: string;
  specialty: string;
  date: string;
  time: string;
  reason?: string;
  status: AppointmentStatus;
  fee: number;
}

export interface CreateAppointmentBody {
  doctorId: number;
  date: string;
  time: string;
  reason?: string;
}

export interface MedicalHistory {
  id: number;
  doctorName: string;
  specialty: string;
  date: string;
  diagnosis: string;
  prescription?: string;
  notes?: string;
}

export interface DashboardSummary {
  totalAppointments: number;
  upcomingAppointments: number;
  completedAppointments: number;
  cancelledAppointments: number;
  totalSpent: number;
  nextAppointment?: Appointment;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  patient: "patient",
  doctor: "doctor",
  receptionist: "receptionist",
  admin: "admin",
} as const;

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  role: AuthUserRole;
  createdAt: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  patient: "patient",
  doctor: "doctor",
  receptionist: "receptionist",
  admin: "admin",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type ListDoctorsParams = {
  specialty?: string;
};

export type ListAppointmentsParams = {
  status?: ListAppointmentsStatus;
};

export type ListAppointmentsStatus =
  (typeof ListAppointmentsStatus)[keyof typeof ListAppointmentsStatus];

export const ListAppointmentsStatus = {
  upcoming: "upcoming",
  completed: "completed",
  cancelled: "cancelled",
} as const;