# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Clinic API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Register a new user
 */
export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  name: zod.string(),
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
  role: zod.enum(["patient", "doctor", "receptionist", "admin"]).optional(),
});

/**
 * @summary Log in
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  role: zod.enum(["patient", "doctor", "receptionist", "admin"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  role: zod.enum(["patient", "doctor", "receptionist", "admin"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List available specialties
 */
export const ListSpecialtiesResponseItem = zod.string();
export const ListSpecialtiesResponse = zod.array(ListSpecialtiesResponseItem);

/**
 * @summary List all doctors
 */
export const ListDoctorsQueryParams = zod.object({
  specialty: zod.coerce.string().optional(),
});

export const ListDoctorsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  specialty: zod.string(),
  qualification: zod.string(),
  experience: zod.number().describe("Years of experience"),
  fee: zod.number().describe("Consultation fee in PKR"),
  availableTimes: zod.array(zod.string()),
  rating: zod.number(),
  totalPatients: zod.number(),
});
export const ListDoctorsResponse = zod.array(ListDoctorsResponseItem);

/**
 * @summary Get a doctor by ID
 */
export const GetDoctorParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDoctorResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  specialty: zod.string(),
  qualification: zod.string(),
  experience: zod.number().describe("Years of experience"),
  fee: zod.number().describe("Consultation fee in PKR"),
  availableTimes: zod.array(zod.string()),
  rating: zod.number(),
  totalPatients: zod.number(),
});

/**
 * @summary List appointments
 */
export const ListAppointmentsQueryParams = zod.object({
  status: zod.enum(["upcoming", "completed", "cancelled"]).optional(),
});

export const ListAppointmentsResponseItem = zod.object({
  id: zod.number(),
  doctorId: zod.number(),
  doctorName: zod.string(),
  specialty: zod.string(),
  date: zod.coerce.date(),
  time: zod.string(),
  reason: zod.string().optional(),
  status: zod.enum(["upcoming", "completed", "cancelled"]),
  fee: zod.number(),
});
export const ListAppointmentsResponse = zod.array(ListAppointmentsResponseItem);

/**
 * @summary Book an appointment
 */
export const CreateAppointmentBody = zod.object({
  doctorId: zod.number(),
  date: zod.coerce.date(),
  time: zod.string(),
  reason: zod.string().optional(),
});

/**
 * @summary Get appointment by ID
 */
export const GetAppointmentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAppointmentResponse = zod.object({
  id: zod.number(),
  doctorId: zod.number(),
  doctorName: zod.string(),
  specialty: zod.string(),
  date: zod.coerce.date(),
  time: zod.string(),
  reason: zod.string().optional(),
  status: zod.enum(["upcoming", "completed", "cancelled"]),
  fee: zod.number(),
});

/**
 * @summary Cancel an appointment
 */
export const CancelAppointmentParams = zod.object({
  id: zod.coerce.number(),
});

export const CancelAppointmentResponse = zod.object({
  id: zod.number(),
  doctorId: zod.number(),
  doctorName: zod.string(),
  specialty: zod.string(),
  date: zod.coerce.date(),
  time: zod.string(),
  reason: zod.string().optional(),
  status: zod.enum(["upcoming", "completed", "cancelled"]),
  fee: zod.number(),
});

/**
 * @summary List medical history records
 */
export const ListMedicalHistoryResponseItem = zod.object({
  id: zod.number(),
  doctorName: zod.string(),
  specialty: zod.string(),
  date: zod.coerce.date(),
  diagnosis: zod.string(),
  prescription: zod.string().optional(),
  notes: zod.string().optional(),
});
export const ListMedicalHistoryResponse = zod.array(
  ListMedicalHistoryResponseItem,
);

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalAppointments: zod.number(),
  upcomingAppointments: zod.number(),
  completedAppointments: zod.number(),
  cancelledAppointments: zod.number(),
  totalSpent: zod.number(),
  nextAppointment: zod
    .object({
      id: zod.number(),
      doctorId: zod.number(),
      doctorName: zod.string(),
      specialty: zod.string(),
      date: zod.coerce.date(),
      time: zod.string(),
      reason: zod.string().optional(),
      status: zod.enum(["upcoming", "completed", "cancelled"]),
      fee: zod.number(),
    })
    .optional(),
});